move scripts to new location
